requirements.txt
# Framework principal
agno>=2.0.0
anthropic>=0.40.0
mcp>=1.0.0
fastapi[standard]>=0.115.0

# Integração WhatsApp
twilio>=8.0.0
python-whatsapp-bot>=1.0.0

# Banco de dados e cache
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# APIs e requisições
httpx>=0.25.0
aiohttp>=3.9.0

# Processamento de linguagem natural (Google Gemini)
google-generativeai>=0.3.0

# Análise de dados
pandas>=2.1.0
numpy>=1.24.0

# Agendamento e tarefas
apscheduler>=3.10.0
celery>=5.3.0

# Configuração e ambiente
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging e monitoramento
loguru>=0.7.0
sentry-sdk>=1.39.0

# Utilidades
python-dateutil>=2.8.0
pytz>=2023.3

# Testes
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
